pyproject.toml
[tool.poetry]
name = "tfg-aletheia"
version = "0.1.0"
description = ""
authors = ["jvega <jvega@ediversa.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.25"
fastapi = "^0.109.0"
aiomysql = "^0.2.0"
aiosqlite = "^0.19.0"
sqlalchemy-utils = "^0.41.1"
modules = "^1.0.0"
python-json-logger = "^2.0.7"
uvicorn = "^0.27.0.post1"
mypy = "^1.8.0"
pydantic-settings = "^2.1.0"
pydantic = "^2.6.1"
psycopg2-binary = "^2.9.9"
colorlog = "^6.8.2"
dependency-injector-fork = "^4.42.1"
lato = "^0.11.0"
httpx = "^0.27.0"
asyncpg = "^0.29.0"
aiofiles = "^23.2.1"
motor = "^3.4.0"
python-multipart = "^0.0.9"
types-aiofiles = "^23.2.0.20240403"
fractal-specifications = "^3.1.6"
jwt = "^1.3.1"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"
aiosmtplib = "^3.0.1"
tenacity = "^8.5.0"
circuitbreaker = "^2.0.0"
aiosmtpd = "^1.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
